feat: show a shinny error on duplicated account
when an user attempt to link a discord account which has already be linked to another SeAT user, it's throwing an exception due to the table integrity constraints.

this is introducing a more user-friendly experience by showing them explicitly they're attempting to link their account to another one.

action is logged with further details which can be useful for staff.